refactor: remove push notification logic from actions and integrate permission request in HomedDescription component
