+++ src/app/home/page.tsx
@@ -6,21 +6,9 @@ import TopSatatusBar from "@/widgets/home/TopSatatusBar";
 import InventoryList from "@/entities/room/ui/InventoryList";
 import CoupleGuard from "@/shared/lib/CoupleGuard";
 import HomedDescription from "@/features/description/ui/HomedDescription";
-import { sendNotification } from "../actions";
 
 const HomePage = () => {
   const [showTour, setShowTour] = useState(false);
-
-  useEffect(() => {
-    const isNotFirst = localStorage.getItem("isNotFirst");
-
-    if (isNotFirst !== "true") {
-      console.log("isNotFirst", isNotFirst);
-
-      sendNotification("홈 화면에 추가해주셔서 감사합니다!");
-      localStorage.setItem("isNotFirst", "false");
-    }
-  }, []);
 
   useEffect(() => {
     const isSeen = localStorage.getItem("homeTourSeen");
